Organized user components, removed debug statements
